Richard's updates to tfMRI for GE
The processEPrime_*.py files have changed substantially. I separated out the code that extracts the subframe for each run and the start time to a new function. I also used os.path.join so that these programs don’t give an error if they are run directly from the command line without an explicit path to the EPrime file (previously it was adding ‘/nback-instructions.txt’ to the path, which wasn’t so good if the path was blank).

I don’t think that the other files have changed much if at all. ReadEPrime now searches for ExperimentName instead of Subject (line 23), since one site had Subject in their file path as well as a column header.